build: add compiler options (and fix warnings)

+++ handmade.cpp
@@ -26,8 +26,8 @@ internal void RenderTestGradient(game_offscreen_buffer *Buffer, int BlueOffset,
 
     uint32 *Pixel = (uint32 *)Row;
     for (int X = 0; X < Buffer->Width; X++) {
-      uint8 Blue = (X + BlueOffset);
-      uint8 Green = (Y + GreenOffset);
+      uint8 Blue = (uint8)(X + BlueOffset);
+      uint8 Green = (uint8)(Y + GreenOffset);
 
       *Pixel++ = ((Green << 8) | Blue);
     }
